requirements.txt
requests
pytest
pytest-html
tenacity
python-dotenv
allure-pytest
